[FIX] sale: wrap long line in sale portal
From Sale create a new (or edit) SO. Add a note to the order with a log
line without dash or whitespace (i.e. a url). Click on "Preview".

The sale portal will try to display the long line which will overflow
the standard size of the column creating an ugly overflow with a very
long horizontal scrollbar.

Adding a "break-all" rule to fix the sale portal case.

opw-2088589

closes odoo/odoo#39968

X-original-commit: 28afbaf502db88204bacbb6be96c490ce777208b
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>